Reorganized chess annealing and tests, develop chess annealing generation and advanced permutation
